docs/site/src/pages/guides/comparison/performance/bench/pyproject.toml
[tool.poetry]
name = "bench"
version = "0.1.0"
description = ""
authors = ["Emiliano Bonilla <56323762+emilbon99@users.noreply.github.com>"]
readme = "README.md"
packages = [
	{ include = "bench/**/*.py" }
]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
influxdb-client = "^1.46.0"
synnax = "^0.32.0"
docker = "^7.1.0"
psycopg2-binary = "^2.9.9"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
